new search UI w/ responsive
