+++ front/src/app/components/search/SearchMiniMap.scss
@@ -1,52 +1,108 @@
 @use "react-design-system/src/config/responsive" as *;
 
+.im-search-filter-counter {
+  background: var(--background-action-high-blue-france);
+  color: var(--text-inverted-blue-france);
+  border-radius: 100%;
+  position: relative;
+  top: -.1rem;
+}
 
 .search-map-results {
-  display: none;
-  overflow: hidden;
-  background-image: linear-gradient(0deg, var(--border-default-grey), var(--border-default-grey)), linear-gradient(0deg, var(--border-default-grey), var(--border-default-grey)), linear-gradient(0deg, var(--border-default-grey), var(--border-default-grey)), linear-gradient(0deg, var(--border-default-grey), var(--border-default-grey));
-  @include for-screen-min($bp-lg) {
-    display: block;
+  &__summary {
+    text-align: right;
   }
+
+  &__overlay {
+    position: fixed;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    background-color: rgba(0, 0, 0, 0.5);
+    pointer-events: none;
+    opacity: 0;
+    z-index: 9998;
+    transition: opacity 0.3s ease-in-out;
+  }
+
+  &__inner {
+    position: fixed;
+    top: 0;
+    bottom: 0;
+    left: 60%;
+    right: 0;
+    transform: translateX(100%);
+    transition: transform 0.3s ease-in-out;
+    z-index: 9999;
+  }
+
+  &--opened {
+    .search-map-results__overlay {
+      pointer-events: auto;
+      opacity: 1;
+    }
+
+    .search-map-results__inner {
+      transform: translateX(0);
+    }
+  }
+
+  &__close-button {
+    position: absolute;
+    top: 1rem;
+    right: 1rem;
+    z-index: 9999;
+    background: var(--background-alt-grey);
+  }
+
   &__map {
-    height: 42vh;
-    min-height: 386px;
+    min-height: 100vh;
   }
 
   .leaflet-control-zoom [href] {
     background-image: none;
   }
+
   .leaflet-popup {
     z-index: 9999;
   }
+
   .leaflet-popup-content {
     margin: 0;
+
     .fr-card {
       background-image: none !important;
       width: 100%;
       height: auto;
+
       &__end {
         margin-bottom: auto !important;
       }
+
       &__content {
         padding: 1rem !important;
       }
+
       &__detail {
         margin-top: 0;
         margin-bottom: 0;
       }
+
       &__title {
         a:after {
           bottom: 1rem !important;
         }
       }
     }
-    
+
   }
+
   .leaflet-popup-content-wrapper {
     overflow-y: auto;
     overflow-x: hidden;
   }
+
   .leaflet-popup-close-button {
     z-index: 9999;
   }
