+++ front/src/app/pages/search/SearchPage.scss
@@ -2,29 +2,8 @@
 
 .im-search-page {
   &__form {
-    display: flex;
-    align-items: flex-end;
-
     &--v2 {
       max-width: 100% !important;
-    }
-  }
-
-  &__results-header {
-    width: 100%;
-  }
-
-  &__results-summary {
-    text-align: right;
-  }
-
-  &__search-filters {
-    display: flex;
-    gap: .8rem;
-    flex-direction: column;
-
-    @include for-screen-min($bp-sm) {
-      flex-direction: row;
     }
   }
 }
