+++ front/src/app/pages/search/SearchPage.tsx
@@ -1,19 +1,11 @@
-import { colors, fr } from "@codegouvfr/react-dsfr";
-import Accordion, {
-  type AccordionProps,
-} from "@codegouvfr/react-dsfr/Accordion";
-import { Button, type ButtonProps } from "@codegouvfr/react-dsfr/Button";
-import ButtonsGroup from "@codegouvfr/react-dsfr/ButtonsGroup";
-import Checkbox from "@codegouvfr/react-dsfr/Checkbox";
-import RadioButtons from "@codegouvfr/react-dsfr/RadioButtons";
-import { Select, type SelectProps } from "@codegouvfr/react-dsfr/SelectNext";
-import { equals, includes, keys } from "ramda";
+import { fr } from "@codegouvfr/react-dsfr";
+import { Button } from "@codegouvfr/react-dsfr/Button";
+import { Select } from "@codegouvfr/react-dsfr/SelectNext";
+import { includes, keys } from "ramda";
 import {
   type ElementRef,
-  type ReactNode,
   useCallback,
   useEffect,
-  useLayoutEffect,
   useRef,
   useState,
 } from "react";
@@ -21,33 +13,14 @@ import {
   Loader,
   MainWrapper,
   PageHeader,
-  RichDropdown,
   SectionAccordion,
   SectionHighlight,
   SectionTextEmbed,
-  useLayout,
   useScrollTo,
 } from "react-design-system";
-import {
-  FormProvider,
-  useForm,
-  useFormContext,
-  useWatch,
-} from "react-hook-form";
+import { FormProvider, useForm, useWatch } from "react-hook-form";
 import { useDispatch } from "react-redux";
-import {
-  domElementIds,
-  type FitForDisableWorkerOption,
-  fitForDisabledWorkersPositiveOptions,
-  type LatLonDistance,
-  type NafSectionSuggestion,
-  type RemoteWorkMode,
-  remoteWorkModeLabels,
-  remoteWorkModes,
-  type SearchSortedBy,
-  searchSortedByOptions,
-  type ValueOf,
-} from "shared";
+import { domElementIds, type ValueOf } from "shared";
 import { Breadcrumbs } from "src/app/components/Breadcrumbs";
 import { AppellationAutocomplete } from "src/app/components/forms/autocomplete/AppellationAutocomplete";
 import { PlaceAutocomplete } from "src/app/components/forms/autocomplete/PlaceAutocomplete";
@@ -61,24 +34,27 @@ import {
   type SearchRoute,
   useSearch,
 } from "src/app/hooks/search.hooks";
+import {
+  areValidGeoParams,
+  canSubmitSearch,
+} from "src/app/pages/search/SearchPage.utils";
 import labonneboiteLogoUrl from "src/assets/img/logo-lbb-centered.png";
 import { featureFlagSelectors } from "src/core-logic/domain/featureFlags/featureFlags.selector";
 import { geosearchSlice } from "src/core-logic/domain/geosearch/geosearch.slice";
-import { nafSelectors } from "src/core-logic/domain/naf/naf.selectors";
+import { nafSlice } from "src/core-logic/domain/naf/naf.slice";
 import { searchSelectors } from "src/core-logic/domain/search/search.selectors";
 import type { SearchPageParams } from "src/core-logic/domain/search/search.slice";
 import { useStyles } from "tss-react/dsfr";
 import "./SearchPage.scss";
-import { nafSlice } from "src/core-logic/domain/naf/naf.slice";
 import Styles from "./SearchPage.styles";
 
-const radiusOptions = ["1", "2", "5", "10", "20", "50", "100"].map(
+export const radiusOptions = ["1", "2", "5", "10", "20", "50", "100"].map(
   (distance) => ({
     label: `${distance} km`,
     value: distance,
   }),
 );
-const DEFAULT_DISTANCE_KM = 10;
+export const DEFAULT_DISTANCE_KM = 10;
 
 const getSearchRouteParam = (
   currentKey: keyof SearchPageParams,
@@ -118,21 +94,32 @@ const defaultValuesForExternalSearch: Required<
   ],
 };
 
-const appellationHintText = (useNaturalLanguageForAppellations?: boolean) =>
+export const appellationInputLabel = (
+  useNaturalLanguageForAppellations?: boolean,
+) =>
+  useNaturalLanguageForAppellations
+    ? "Que souhaitez-vous faire ou découvrir ?"
+    : "Je recherche le métier...";
+
+export const appellationHintText = (
+  useNaturalLanguageForAppellations?: boolean,
+) =>
   useNaturalLanguageForAppellations
     ? "Un métier, un domaine ou une activité qui vous intéresse"
     : "Un métier qui vous intéresse";
 
-const appellationPlaceholder = (useNaturalLanguageForAppellations?: boolean) =>
+export const appellationPlaceholder = (
+  useNaturalLanguageForAppellations?: boolean,
+) =>
   useNaturalLanguageForAppellations
     ? "Boulanger, comptabilité, travailler avec des enfants..."
     : "Ex : Boulanger, styliste, etc";
 
-const radiusHintText = "Affinez votre localisation";
+export const radiusHintText = "Affinez votre localisation";
 
-const placeHintText = "Une ville, un département, une région, etc.";
+export const placeHintText = "Une ville, un département, une région, etc.";
 
-const placeInputLabel = <>Dans quelle ville ?</>;
+export const placeInputLabel = <>Dans quelle ville ?</>;
 
 export const SearchPage = ({
   route,
@@ -157,7 +144,6 @@ export const SearchPage = ({
   const { enableSearchByScore } = useAppSelector(
     featureFlagSelectors.featureFlagState,
   );
-  const [isPanelOpened, setIsPanelOpened] = useState(false);
   const initialValues: SearchPageParams = {
     sortBy: enableSearchByScore ? "score" : "date",
     sortOrder: "desc",
@@ -213,7 +199,7 @@ export const SearchPage = ({
     ),
     mode: "onTouched",
   });
-  const { handleSubmit, setValue, register, control, getValues } = methods;
+  const { handleSubmit, setValue, control, getValues, reset } = methods;
   const formValues = getValues();
   const [lat, lon, distanceKm, place] = useWatch({
     control,
@@ -221,8 +207,6 @@ export const SearchPage = ({
   });
 
   const searchHasBeenMade = searchMade !== null;
-
-  const nafOptions = useAppSelector(nafSelectors.allSections);
 
   const internalSearchIsAvailableForInitialSearchRequest =
     !isExternal &&
@@ -240,36 +224,13 @@ export const SearchPage = ({
     distanceKm !== 0;
   routeParams.appellationCodes && routeParams.appellationCodes.length > 0;
 
-  const filteredSortOptions = getSortedByOptions(
-    areValidGeoParams(formValues),
-    enableSearchByScore.isActive,
-  );
-
-  const appellationInputLabel = (
-    <>
-      {useNaturalLanguageForAppellations
-        ? "Que souhaitez-vous faire ou découvrir ?"
-        : "Je recherche le métier..."}
-    </>
-  );
-
-  const setTempValuesAsFormValues = useCallback(
-    (values: Partial<SearchPageParams>) => {
-      keys(values).forEach((key) => {
-        setValue(key, values[key]);
-      });
-    },
-    [setValue],
-  );
-
   const onSearchFormSubmit = useCallback(
     (updatedValues: SearchPageParams) => {
-      setTempValue(updatedValues);
-      setTempValuesAsFormValues(updatedValues);
       setSearchMade(updatedValues);
+      reset(updatedValues);
       triggerSearch(filterFormValues(updatedValues), isExternal);
     },
-    [setTempValuesAsFormValues, triggerSearch, filterFormValues, isExternal],
+    [triggerSearch, filterFormValues, isExternal],
   );
 
   useScrollTo(pagination?.currentPage ?? 1);
@@ -304,21 +265,6 @@ export const SearchPage = ({
     dispatch(nafSlice.actions.getAllSectionsRequested());
   }, [dispatch]);
 
-  const displayTotalFiltersLength = (formValues: SearchPageParams): string => {
-    const currentFiltersLength = [
-      ...getAppellationAndNafValues(
-        formValues.appellationCodes,
-        formValues.nafCodes,
-      ),
-      ...getEstablishmentAdditionalValues(
-        formValues.fitForDisabledWorkers,
-        formValues.showOnlyAvailableOffers,
-      ),
-      ...getLocalisationValues(formValues.place, formValues.remoteWorkModes),
-    ].length;
-    return currentFiltersLength > 0 ? `(${currentFiltersLength})` : "";
-  };
-
   return (
     <HeaderFooterLayout>
       <MainWrapper vSpacing={0} layout="fullscreen">
@@ -347,7 +293,9 @@ export const SearchPage = ({
                 <div className={cx(fr.cx("fr-col-12", "fr-col-lg-4"))}>
                   <AppellationAutocomplete
                     locator="search-form-appellation"
-                    label={appellationInputLabel}
+                    label={appellationInputLabel(
+                      useNaturalLanguageForAppellations,
+                    )}
                     hintText={appellationHintText(
                       useNaturalLanguageForAppellations,
                     )}
@@ -502,692 +450,34 @@ export const SearchPage = ({
                     onSearchFormSubmit(updatedValues);
                   }
                 })}
-                className={cx(
-                  fr.cx("fr-container", "fr-mb-4w"),
-                  Styles.searchFilters,
-                )}
+                className={cx(Styles.searchFilters)}
                 id={domElementIds[route.name].searchForm}
               >
-                <SearchFilterPanel
-                  initialValues={initialValues}
-                  tempValue={tempValue}
-                  appellationInputLabel={appellationInputLabel}
-                  useNaturalLanguageForAppellations={
-                    useNaturalLanguageForAppellations
-                  }
-                  routeName={route.name}
-                  isPanelOpened={isPanelOpened}
-                  setIsPanelOpened={setIsPanelOpened}
-                  appellationHintText={appellationHintText}
-                  onSearchFormSubmit={onSearchFormSubmit}
-                  setTempValue={setTempValue}
-                  nafOptions={nafOptions}
-                />
-                <section
-                  className={cx(
-                    fr.cx("fr-grid-row", "fr-grid-row--middle"),
-                    Styles.resultsHeader,
+                <div ref={searchResultsWrapper}>
+                  {searchMade !== null && (
+                    <div
+                      ref={innerSearchResultWrapper}
+                      className={fr.cx("fr-pb-1w")}
+                    >
+                      <SearchListResults
+                        route={route}
+                        showDistance={areValidGeoParams(searchMade)}
+                        isExternal={route.name === "externalSearch"}
+                        onSearchFormSubmit={onSearchFormSubmit}
+                        useNaturalLanguageForAppellations={
+                          useNaturalLanguageForAppellations
+                        }
+                        tempValue={tempValue}
+                        setTempValue={setTempValue}
+                      />
+                    </div>
                   )}
-                >
-                  <Button
-                    type="button"
-                    priority="secondary"
-                    iconId="fr-icon-equalizer-line"
-                    size="small"
-                    onClick={() => setIsPanelOpened(true)}
-                  >
-                    Filtrer {displayTotalFiltersLength(formValues)}
-                  </Button>
-                  <div className={fr.cx("fr-hidden", "fr-unhidden-lg")}>
-                    <RichDropdown
-                      defaultValue="Trier par pertinence"
-                      iconId="fr-icon-arrow-down-line"
-                      id={domElementIds[route.name].sortFilterTag}
-                      values={
-                        formValues.sortBy
-                          ? [sortedByOptionsLabel[formValues.sortBy]]
-                          : []
-                      }
-                      className={fr.cx("fr-ml-2w")}
-                      submenu={{
-                        title: "Ordre d’affichage",
-                        content: (
-                          <RadioButtons
-                            id={domElementIds[route.name].sortRadioButtons}
-                            options={filteredSortOptions.map((option) => ({
-                              ...option,
-                              nativeInputProps: {
-                                name: register("sortBy").name,
-                                value: option.value,
-                                disabled: option.disabled,
-                                checked: tempValue
-                                  ? option.value === tempValue.sortBy
-                                  : false,
-                                onClick: (event) => {
-                                  const updatedSortedBy = isSearchSortedBy(
-                                    event.currentTarget.value,
-                                  )
-                                    ? event.currentTarget.value
-                                    : "score";
-                                  if (
-                                    updatedSortedBy === "distance" &&
-                                    areValidGeoParams(tempValue)
-                                  ) {
-                                    setTempValue({
-                                      ...tempValue,
-                                      sortBy: "distance",
-                                    });
-                                  }
-                                  if (updatedSortedBy !== "distance") {
-                                    setTempValue({
-                                      ...tempValue,
-                                      sortBy: updatedSortedBy,
-                                    });
-                                  }
-                                },
-                              },
-                            }))}
-                          />
-                        ),
-                      }}
-                    />
-                  </div>
-
-                  <div
-                    className={cx(fr.cx("fr-ml-auto"), Styles.resultsSummary)}
-                  >
-                    {!isExternal && (
-                      <>
-                        <h2 className={fr.cx("fr-h5", "fr-mb-0")}>
-                          <strong>{pagination.totalRecords}</strong> résultat
-                          {pagination.totalRecords > 1 ? "s" : ""} trouvé
-                          {pagination.totalRecords > 1 ? "s" : ""}
-                        </h2>
-                        {routeParams.appellations &&
-                          routeParams.appellations.length > 0 && (
-                            <span className={cx(fr.cx("fr-text--xs"))}>
-                              pour la recherche{" "}
-                              <a
-                                href={`https://candidat.francetravail.fr/metierscope/fiche-metier/${routeParams.appellations[0].romeCode}`}
-                                target="_blank"
-                                className={fr.cx("fr-text--bold")}
-                                rel="noreferrer"
-                              >
-                                {routeParams.appellations[0].appellationLabel}
-                              </a>
-                            </span>
-                          )}
-                      </>
-                    )}
-                  </div>
-                </section>
+                </div>
               </form>
             </FormProvider>
-
-            <div ref={searchResultsWrapper}>
-              {searchMade !== null && (
-                <div
-                  ref={innerSearchResultWrapper}
-                  className={fr.cx("fr-pb-1w")}
-                >
-                  <SearchListResults
-                    route={route}
-                    showDistance={areValidGeoParams(searchMade)}
-                    isExternal={route.name === "externalSearch"}
-                    onSearchFormSubmit={onSearchFormSubmit}
-                  />
-                </div>
-              )}
-            </div>
           </>
         )}
       </MainWrapper>
     </HeaderFooterLayout>
-  );
-};
-
-const areValidGeoParams = (
-  geoParams: Partial<LatLonDistance>,
-): geoParams is LatLonDistance => {
-  return (
-    geoParams.latitude !== undefined &&
-    geoParams.longitude !== undefined &&
-    geoParams.distanceKm !== undefined &&
-    geoParams.distanceKm > 0
-  );
-};
-
-const areEmptyGeoParams = (
-  geoParams: Partial<LatLonDistance>,
-): geoParams is Partial<LatLonDistance> => {
-  return (
-    geoParams.latitude === undefined &&
-    geoParams.longitude === undefined &&
-    geoParams.distanceKm === undefined
-  );
-};
-
-const canSubmitSearch = (values: SearchPageParams) => {
-  const geoParams = {
-    latitude: values.latitude,
-    longitude: values.longitude,
-    distanceKm: values.distanceKm,
-  };
-  return areValidGeoParams(geoParams) || areEmptyGeoParams(geoParams);
-};
-
-export const getSortedByOptions = (
-  hasGeoParams: boolean,
-  hasScoreEnabled: boolean,
-): SelectProps.Option<SearchSortedBy>[] => [
-  ...(hasScoreEnabled
-    ? [
-        {
-          label: sortedByOptionsLabel.score,
-          value: "score" as const,
-        },
-      ]
-    : []),
-  {
-    label: sortedByOptionsLabel.date,
-    value: "date" as const,
-  },
-  {
-    label: sortedByOptionsLabel.distance,
-    value: "distance" as const,
-    disabled: !hasGeoParams,
-  },
-];
-
-const sortedByOptionsLabel = {
-  date: "Trier par date de publication",
-  score: "Trier par pertinence",
-  distance: "Trier par proximité",
-};
-
-const isSearchSortedBy = (value: string): value is SearchSortedBy =>
-  searchSortedByOptions.includes(value as SearchSortedBy);
-
-const rqthLabel = "Personnes en situation de handicap bienvenues";
-
-type FilterAccordionProps = AccordionProps & {
-  onReset?: () => void;
-  values: string[];
-};
-
-type SearchFilterPanelProps = {
-  initialValues: SearchPageParams;
-  tempValue: SearchPageParams;
-  appellationInputLabel: ReactNode;
-  useNaturalLanguageForAppellations?: boolean;
-  routeName: SearchRoute["name"];
-  isPanelOpened: boolean;
-  setIsPanelOpened: (isOpened: boolean) => void;
-  appellationHintText: (useNaturalLanguageForAppellations?: boolean) => string;
-  onSearchFormSubmit: (updatedValues: SearchPageParams) => void;
-  setTempValue: (updatedValues: SearchPageParams) => void;
-  nafOptions: NafSectionSuggestion[];
-};
-
-const getLocalisationValues = (
-  place?: string,
-  currentRemoteWorkModes?: RemoteWorkMode[],
-): string[] => [
-  ...(place ? [place] : []),
-  ...((currentRemoteWorkModes || []).length === remoteWorkModes.length
-    ? []
-    : currentRemoteWorkModes || []),
-];
-
-const getEstablishmentAdditionalValues = (
-  fitForDisabledWorkers?: FitForDisableWorkerOption[],
-  showOnlyAvailableOffers?: boolean,
-): string[] => [
-  ...(equals(fitForDisabledWorkers, [...fitForDisabledWorkersPositiveOptions])
-    ? [(fitForDisabledWorkers || []).join(", ")]
-    : []),
-  ...(showOnlyAvailableOffers ? [showOnlyAvailableOffers.toString()] : []),
-];
-
-const getAppellationAndNafValues = (
-  appellationCodes?: string[],
-  nafCodes?: string[],
-): string[] => [
-  ...(appellationCodes ? [appellationCodes.join(", ")] : []),
-  ...(nafCodes ? [nafCodes.join(", ")] : []),
-];
-
-const SearchFilterPanel = ({
-  initialValues,
-  tempValue,
-  appellationInputLabel,
-  useNaturalLanguageForAppellations,
-  routeName,
-  isPanelOpened,
-  setIsPanelOpened,
-  appellationHintText,
-  onSearchFormSubmit,
-  setTempValue,
-  nafOptions,
-}: SearchFilterPanelProps) => {
-  const {
-    clearErrors,
-    getValues,
-    register,
-    formState: { errors },
-    watch,
-  } = useFormContext<SearchPageParams>();
-  const { isLayoutDesktop } = useLayout();
-  const formValues = getValues();
-  const [place] = watch(["place"]);
-  const wrapperElementRef = useRef<HTMLDivElement>(null);
-  const filtersSectionMaxWidth = isLayoutDesktop ? 420 : "100%";
-
-  useLayoutEffect(() => {
-    const frHeader = document.getElementById("fr-header");
-    const refreshTopOffset = () => {
-      if (wrapperElementRef.current) {
-        wrapperElementRef.current.style.top = `${
-          frHeader && frHeader.offsetHeight - window.scrollY > 0
-            ? frHeader.offsetHeight - window.scrollY
-            : 0
-        }px`;
-      }
-    };
-
-    refreshTopOffset();
-    window.addEventListener("scroll", refreshTopOffset);
-    window.addEventListener("resize", refreshTopOffset);
-
-    return () => {
-      window.removeEventListener("scroll", refreshTopOffset);
-      window.removeEventListener("resize", refreshTopOffset);
-    };
-  }, []);
-
-  return (
-    <section
-      ref={wrapperElementRef}
-      style={{
-        maxWidth: filtersSectionMaxWidth,
-        position: "fixed",
-        left: 0,
-        top: 0,
-        ...(isLayoutDesktop ? {} : { right: 0 }),
-        bottom: 0,
-        zIndex: 1000,
-        boxShadow: "0 5px 5px 0 rgba(0,0,0,0.5)",
-        backgroundColor: colors.decisions.background.default.grey.default,
-        transform: `translateX(${isPanelOpened ? 0 : -100}%)`,
-        transition: "transform ease .2s",
-      }}
-      className={fr.cx("fr-pt-6w", "fr-px-2w", "fr-pb-2w")}
-    >
-      <Button
-        style={{ position: "absolute", top: 0, right: 0 }}
-        className={fr.cx("fr-btn--close", "fr-mt-1w", "fr-mr-1w")}
-        title="Fermer"
-        aria-controls="shareLink"
-        type="button"
-        data-fr-js-modal-button="true"
-        onClick={() => setIsPanelOpened(false)}
-      >
-        Fermer
-      </Button>
-      <h3 className={fr.cx("fr-h5")}>Affiner la recherche</h3>
-      <div className={fr.cx("fr-accordions-group")}>
-        <FilterAccordion
-          label="Métier ou secteur d'activité"
-          values={getAppellationAndNafValues(
-            formValues.appellationCodes,
-            formValues.nafCodes,
-          )}
-          defaultExpanded={true}
-          onReset={() => {
-            const updatedValues = {
-              ...tempValue,
-              appellations: undefined,
-              appellationCodes: undefined,
-            };
-            onSearchFormSubmit(updatedValues);
-          }}
-        >
-          <AppellationAutocomplete
-            locator="search-form-appellation"
-            className={fr.cx("fr-mb-2w")}
-            label={appellationInputLabel}
-            hintText={appellationHintText(useNaturalLanguageForAppellations)}
-            onAppellationSelected={(appellationMatch) => {
-              clearErrors("appellations");
-              setTempValue({
-                ...tempValue,
-                appellations: [appellationMatch.appellation],
-                appellationCodes: [
-                  appellationMatch.appellation.appellationCode,
-                ],
-              });
-            }}
-            onAppellationClear={() => {
-              setTempValue({
-                ...tempValue,
-                appellations: undefined,
-                appellationCodes: undefined,
-              });
-            }}
-            selectProps={{
-              inputId: domElementIds[routeName].appellationAutocomplete,
-              placeholder: appellationPlaceholder(
-                useNaturalLanguageForAppellations,
-              ),
-              defaultValue: formValues.appellations?.[0]
-                ? {
-                    label: formValues.appellations?.[0]?.appellationLabel,
-                    value: {
-                      appellation: formValues.appellations?.[0],
-                      matchRanges: [],
-                    },
-                  }
-                : undefined,
-            }}
-            state={errors.appellations ? "error" : undefined}
-            stateRelatedMessage={errors.appellations?.message}
-          />
-          {tempValue.appellations?.length && (
-            <p className={fr.cx("fr-hint-text", "fr-mt-2w")}>
-              <strong>
-                Les résultats seront étendus aux autres métiers du secteur "
-                {tempValue.appellations[0].romeLabel}"
-              </strong>
-              , c’est pour cela que vous pourrez voir des métiers proches mais
-              ne correspondant pas précisément à votre recherche dans les
-              résultats.
-            </p>
-          )}
-          <Select
-            label="Et / ou un secteur d'activité"
-            options={nafOptions.map((option) => ({
-              label: option.label,
-              value: JSON.stringify({
-                nafLabel: option.label,
-                nafCodes: option.nafCodes,
-              }),
-            }))}
-            nativeSelectProps={{
-              id: domElementIds[routeName].nafAutocomplete,
-              value:
-                tempValue.nafLabel && tempValue.nafCodes
-                  ? JSON.stringify({
-                      nafLabel: tempValue.nafLabel,
-                      nafCodes: tempValue.nafCodes,
-                    })
-                  : "",
-              onChange: (event) => {
-                const value = event.currentTarget.value;
-                if (!value) {
-                  setTempValue({
-                    ...tempValue,
-                    nafCodes: undefined,
-                    nafLabel: undefined,
-                  });
-                  return;
-                }
-                const { nafLabel, nafCodes } = JSON.parse(value);
-                setTempValue({
-                  ...tempValue,
-                  nafCodes,
-                  nafLabel,
-                });
-              },
-            }}
-          />
-        </FilterAccordion>
-        <FilterAccordion
-          label="Localisation"
-          values={getLocalisationValues(
-            formValues.place,
-            formValues.remoteWorkModes,
-          )}
-          onReset={() => {
-            const updatedValues: SearchPageParams =
-              tempValue.sortBy === "distance"
-                ? {
-                    ...tempValue,
-                    place: initialValues.place,
-                    latitude: 0,
-                    longitude: 0,
-                    distanceKm: 0,
-                  }
-                : {
-                    ...tempValue,
-                    place: initialValues.place,
-                    latitude: initialValues.latitude,
-                    longitude: initialValues.longitude,
-                    distanceKm: undefined,
-                  };
-            onSearchFormSubmit(updatedValues);
-          }}
-        >
-          <PlaceAutocomplete
-            locator="search-form-place"
-            label={placeInputLabel}
-            hintText={placeHintText}
-            onPlaceSelected={(lookupSearchResult) => {
-              clearErrors("place");
-              if (!lookupSearchResult) return;
-              const newValues = {
-                place: lookupSearchResult.label,
-                latitude: lookupSearchResult.position.lat,
-                longitude: lookupSearchResult.position.lon,
-              };
-              setTempValue({
-                ...tempValue,
-                ...newValues,
-                distanceKm: tempValue.distanceKm || DEFAULT_DISTANCE_KM,
-              });
-            }}
-            onPlaceClear={() => {
-              const updatedInitialValues: SearchPageParams =
-                tempValue.sortBy === "distance"
-                  ? {
-                      ...tempValue,
-                      place: initialValues.place,
-                      latitude: 0,
-                      longitude: 0,
-                      distanceKm: 0,
-                    }
-                  : {
-                      ...tempValue,
-                      place: initialValues.place,
-                      latitude: initialValues.latitude,
-                      longitude: initialValues.longitude,
-                      distanceKm: undefined,
-                    };
-              setTempValue(updatedInitialValues);
-
-              if (formValues.sortBy === "distance") {
-                setTempValue({
-                  ...tempValue,
-                  sortBy: "date",
-                });
-              }
-            }}
-            className={fr.cx("fr-mt-2w")}
-            initialInputValue={place}
-            selectProps={{
-              inputId: domElementIds[routeName].placeAutocompleteInput,
-              defaultValue:
-                place && formValues.latitude && formValues.longitude
-                  ? {
-                      label: place,
-                      value: {
-                        label: place,
-                        position: {
-                          lat: formValues.latitude,
-                          lon: formValues.longitude,
-                        },
-                      },
-                    }
-                  : undefined,
-            }}
-            state={errors.place ? "error" : undefined}
-            stateRelatedMessage={errors.place?.message}
-          />
-
-          <Select
-            label="Dans un rayon de :"
-            options={radiusOptions}
-            disabled={!tempValue.latitude || !tempValue.longitude}
-            hint={radiusHintText}
-            nativeSelectProps={{
-              ...register("distanceKm"),
-              title:
-                !tempValue.latitude || !tempValue.longitude
-                  ? "Pour sélectionner une distance, vous devez d'abord définir une ville."
-                  : undefined,
-              id: domElementIds[routeName].distanceSelect,
-              value: `${tempValue.distanceKm || ""}`,
-              onChange: (event) => {
-                const value = Number.parseInt(event.currentTarget.value, 10);
-                setTempValue({
-                  ...tempValue,
-                  distanceKm: value,
-                });
-                if (!value) {
-                  setTempValue({
-                    ...tempValue,
-                    distanceKm: value,
-                  });
-                }
-              },
-            }}
-          />
-          <p className={fr.cx("fr-text--bold", "fr-mt-2w")}>
-            Quelle(s) offre(s) souhaitez-vous voir ?
-          </p>
-          <Checkbox
-            options={remoteWorkModes.map((mode) => ({
-              label: remoteWorkModeLabels[mode].label,
-              nativeInputProps: {
-                checked: tempValue.remoteWorkModes?.includes(mode),
-                onChange: () => {
-                  const remoteWorkModesArray = tempValue.remoteWorkModes || [];
-                  setTempValue({
-                    ...tempValue,
-                    remoteWorkModes: tempValue.remoteWorkModes?.includes(mode)
-                      ? remoteWorkModesArray.filter((m) => m !== mode)
-                      : [...remoteWorkModesArray, mode],
-                  });
-                },
-              },
-            }))}
-          />
-        </FilterAccordion>
-        <FilterAccordion
-          label="Caractéristiques de l'entreprise"
-          values={getEstablishmentAdditionalValues(
-            formValues.fitForDisabledWorkers,
-            formValues.showOnlyAvailableOffers,
-          )}
-          onReset={() => {
-            const updatedValues = {
-              ...tempValue,
-              fitForDisabledWorkers: undefined,
-            };
-            onSearchFormSubmit(updatedValues);
-          }}
-        >
-          <p className={fr.cx("fr-mb-2w")}>
-            Afficher uniquement les entreprises&nbsp;:
-          </p>
-          <Checkbox
-            className={fr.cx("fr-mb-2w")}
-            options={[
-              {
-                label: rqthLabel,
-                nativeInputProps: {
-                  checked:
-                    tempValue.fitForDisabledWorkers?.some(
-                      (fitForDisabledWorker) =>
-                        fitForDisabledWorker === "yes-declared-only" ||
-                        fitForDisabledWorker === "yes-ft-certified",
-                    ) ?? false,
-                  onChange: (event) => {
-                    setTempValue({
-                      ...tempValue,
-                      fitForDisabledWorkers: event.currentTarget.checked
-                        ? [...fitForDisabledWorkersPositiveOptions]
-                        : ["no"],
-                    });
-                  },
-                },
-              },
-            ]}
-          />
-          <Checkbox
-            className={fr.cx("fr-mb-2w")}
-            options={[
-              {
-                label: "Mises en relation disponibles",
-                nativeInputProps: {
-                  checked: tempValue.showOnlyAvailableOffers,
-                  onChange: (event) => {
-                    setTempValue({
-                      ...tempValue,
-                      showOnlyAvailableOffers: event.currentTarget.checked,
-                    });
-                  },
-                },
-              },
-            ]}
-          />
-        </FilterAccordion>
-      </div>
-    </section>
-  );
-};
-
-const FilterAccordion = ({
-  children,
-  onReset,
-  values,
-  ...accordionProps
-}: FilterAccordionProps) => {
-  const countValues = values?.length || 0;
-  const hasValue = countValues > 0;
-
-  const buttons: [ButtonProps, ...ButtonProps[]] = [
-    {
-      children: "Appliquer",
-      type: "submit",
-      className: fr.cx("fr-mb-0"),
-      id: `${accordionProps.id}-submit-button`,
-    },
-  ];
-  if (onReset) {
-    buttons.unshift({
-      children: "Réinitialiser",
-      type: "button",
-      priority: "tertiary",
-      onClick: () => {
-        onReset();
-      },
-      className: fr.cx("fr-mb-0"),
-      disabled: !hasValue,
-      id: `${accordionProps.id}-reset-button`,
-    });
-  }
-  return (
-    <Accordion
-      {...accordionProps}
-      label={accordionProps.label + (countValues ? ` (${countValues})` : "")}
-    >
-      {children}
-      <ButtonsGroup
-        className={fr.cx("fr-hr", "fr-pt-2w", "fr-pb-0")}
-        inlineLayoutWhen="always"
-        alignment="right"
-        buttons={buttons}
-      />
-    </Accordion>
   );
 };
